a few more updates
summarize_each_group was dropped from pericircle and doesnt have a permanent home at the moment. It may be best as a helper function that pericircumference lets you modify as needed. Default dictionaries are a new part of perinary and they should absolutely be used at the beginning of rmd docs. If we start holding off on recoding until the very last moment, this format would work very nicely